search-service/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
elasticsearch==8.17.2
requests==2.32.3
python-dotenv==1.1.0
pydantic==2.12.5
urllib3==2.4.0
